refactor: consolidate API endpoints
