+++ api/probe-engine/linkTraceLib.js
@@ -2,7 +2,7 @@
 import fetch from 'node-fetch';
 import { URL } from 'url';
 
-export default async function handler(req, res) {
+export async function linkTraceHandler(req, res) {
   // Set CORS headers
   res.setHeader('Access-Control-Allow-Origin', '*');
   res.setHeader('Access-Control-Allow-Methods', 'POST, OPTIONS');
