+++ api/probe-engine/puppeteerProbeLib.js
@@ -448,7 +448,7 @@ async function traceDeviceScenario(url, scenarioId, config, maxHops = 20) {
 }
 
 // Main handler function
-export default async function handler(req, res) {
+export async function puppeteerProbeHandler(req, res) {
   try {
     // Set CORS headers
     res.setHeader('Access-Control-Allow-Origin', '*');
